LayoutTests/fullscreen/full-screen-zIndex-after-expected.txt

This tests that an element with a positive z-index appears behind the full screen element. After entering full screen mode, the green box should not be obscured by the red box. Click go full screen to run the test.
EVENT(webkitfullscreenchange)
EVENT(webkitfullscreenchange)
END OF TEST
